+++ JiraGitHubPRCreator.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 14
 VisualStudioVersion = 14.0.23107.0
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "JiraGitHubPRCreator", "JiraGitHubPRCreator.csproj", "{10020C97-1F30-433B-B46F-1E482C76D79F}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "JiraGitHubPRCreator", "JiraGitHubPRCreator\JiraGitHubPRCreator.csproj", "{10020C97-1F30-433B-B46F-1E482C76D79F}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "JiraGitHubPRCreator.Core", "JiraGitHubPRCreator.Core\JiraGitHubPRCreator.Core.csproj", "{B71D8958-82E7-4F17-AFE6-08D8CA2C99F4}"
 EndProject
